requirements.txt
dj-database-url==2.1.0
Django==5.0.1
django-bootstrap4==23.4
django-filter==24.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
geopy==1.12.0
Markdown==2.6.11
psycopg2==2.9.9 --no-binary=psycopg2
pytz==2024.1
django-phonenumber-field==7.3.0
django-map-widgets==0.4.2
pytest==3.5.0
pytest-django==3.1.2
factory-boy==2.10.0
drf-writable-nested==0.7.0
googlemaps==2.5.1
django-bootstrap-datepicker-plus==2.0.0
django-datetime-widget==0.9.3
drf-dynamic-fields==0.3.0
django-impersonate==1.9.2
channels==4.0.0
channels-redis==4.2.0
africastalking==1.1.4
pytest-mock==1.10.0
django-push-notifications==3.0.2
Pillow==10.2.0
django-extra-fields==3.0.2
django-inline-svg==0.1.1
django-file-form==3.6.0
django-select2==8.1.2
